ESP32-S3/MICS3/test/requirements.txt
action-msgs==2.2.0
action-tutorials-interfaces==0.34.2
action-tutorials-py==0.35.1
actionlib-msgs==5.4.2
ament-cmake-test==2.7.2
ament-copyright==0.19.0
ament-cppcheck==0.19.0
ament-cpplint==0.19.0
ament-flake8==0.19.0
ament-index-python==1.10.1
ament-lint==0.19.0
ament-lint-cmake==0.19.0
ament-package==0.17.1
ament-pep257==0.19.0
ament-uncrustify==0.19.0
ament-xmllint==0.19.0
angles==1.16.0
builtin-interfaces==2.2.0
composition-interfaces==2.2.0
cv-bridge==4.1.0
demo-nodes-py==0.35.1
diagnostic-msgs==5.4.2
domain-coordinator==0.13.1
example-interfaces==0.13.0
examples-rclpy-executors==0.20.3
examples-rclpy-minimal-action-client==0.20.3
examples-rclpy-minimal-action-server==0.20.3
examples-rclpy-minimal-client==0.20.3
examples-rclpy-minimal-publisher==0.20.3
examples-rclpy-minimal-service==0.20.3
examples-rclpy-minimal-subscriber==0.20.3
geometry-msgs==5.4.2
image-geometry==4.1.0
interactive-markers==2.6.1
laser-geometry==2.9.0
launch==3.7.0
launch-ros==0.28.0
launch-testing==3.7.0
launch-testing-ros==0.28.0
launch-xml==3.7.0
launch-yaml==3.7.0
lifecycle-msgs==2.2.0
logging-demo==0.35.1
map-msgs==2.5.0
message-filters==6.0.7
nav-msgs==5.4.2
numpy==2.2.0
osrf-pycommon==2.1.4
pcl-msgs==1.0.0
pendulum-msgs==0.35.1
PyAudio==0.2.14
pyserial==3.5
python-qt-binding==2.3.1
qt-dotgraph==2.8.3
qt-gui==2.8.3
qt-gui-cpp==2.8.3
qt-gui-py-common==2.8.3
quality-of-service-demo-py==0.35.1
rcl-interfaces==2.2.0
rclpy==7.7.0
rcutils==6.9.3
resource-retriever==3.6.0
rmw-dds-common==3.2.0
ros2action==0.36.0
ros2bag==0.29.0
ros2bag-mcap-cli==0.29.0
ros2bag-sqlite3-cli==0.29.0
ros2cli==0.36.0
ros2component==0.36.0
ros2doctor==0.36.0
ros2interface==0.36.0
ros2launch==0.28.0
ros2lifecycle==0.36.0
ros2multicast==0.36.0
ros2node==0.36.0
ros2param==0.36.0
ros2pkg==0.36.0
ros2run==0.36.0
ros2service==0.36.0
ros2topic==0.36.0
rosbag2-interfaces==0.29.0
rosbag2-py==0.29.0
rosgraph-msgs==2.2.0
rosidl-adapter==4.9.2
rosidl-cli==4.9.2
rosidl-cmake==4.9.2
rosidl-generator-c==4.9.2
rosidl-generator-cpp==4.9.2
rosidl-generator-py==0.24.0
rosidl-generator-type-description==4.9.2
rosidl-parser==4.9.2
rosidl-pycommon==4.9.2
rosidl-runtime-py==0.14.0
rosidl-typesupport-c==3.3.1
rosidl-typesupport-cpp==3.3.1
rosidl-typesupport-fastrtps-c==3.7.1
rosidl-typesupport-fastrtps-cpp==3.7.1
rosidl-typesupport-introspection-c==4.9.2
rosidl-typesupport-introspection-cpp==4.9.2
rpyutils==0.6.0
rqt-action==2.3.0
rqt-bag==2.0.1
rqt-bag-plugins==2.0.1
rqt-console==2.3.0
rqt-graph==1.6.1
rqt-gui==1.8.0
rqt-gui-py==1.8.0
rqt-msg==1.6.0
rqt-plot==1.5.0
rqt-publisher==1.8.0
rqt-py-common==1.8.0
rqt-py-console==1.3.0
rqt-reconfigure==1.7.0
rqt-service-caller==1.3.0
rqt-shell==1.3.0
rqt-srv==1.3.0
rqt-topic==1.8.0
sensor-msgs==5.4.2
sensor-msgs-py==5.4.2
service-msgs==2.2.0
shape-msgs==5.4.2
sros2==0.15.1
statistics-msgs==2.2.0
std-msgs==5.4.2
std-srvs==5.4.2
stereo-msgs==5.4.2
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.39.3
tf2-kdl==0.39.3
tf2-msgs==0.39.3
tf2-py==0.39.3
tf2-ros-py==0.39.3
tf2-sensor-msgs==0.39.3
tf2-tools==0.39.3
topic-monitor==0.35.1
trajectory-msgs==5.4.2
turtlesim-msgs==1.9.2
type-description-interfaces==2.2.0
unique-identifier-msgs==2.7.0
visualization-msgs==5.4.2
